fix: Cast SUM results to BIGINT to match function return type
- Added explicit BIGINT casting for total_click_count and unique_visitor_count
- Fixes type mismatch error where SUM returns numeric but function expects bigint
- Ensures proper type compatibility with function signature

+++ supabase/migrations/20250114181000_fix_ambiguous_column.sql
@@ -41,8 +41,8 @@ BEGIN
       sl.destination_url,
       array_agg(ds.date ORDER BY ds.date) as dates,
       array_agg(COALESCE(ldc.daily_clicks, 0) ORDER BY ds.date) as clicks,
-      COALESCE(SUM(ldc.daily_clicks), 0) as total_click_count,
-      COALESCE(SUM(ldc.daily_unique), 0) as unique_visitor_count
+      COALESCE(SUM(ldc.daily_clicks), 0)::BIGINT as total_click_count,
+      COALESCE(SUM(ldc.daily_unique), 0)::BIGINT as unique_visitor_count
     FROM short_links sl
     CROSS JOIN date_series ds
     LEFT JOIN link_daily_clicks ldc ON
